refactor: separating hook
