+++ packages/sanity/src/core/releases/tool/detail/ReleaseDashboardFooter.tsx
@@ -4,7 +4,7 @@ import {useMemo} from 'react'
 
 import {isReleaseScheduledOrScheduling, type ReleaseDocument} from '../../index'
 import {ReleasePublishAllButton} from '../components/releaseCTAButtons/ReleasePublishAllButton'
-import {ReleaseRevertButton} from '../components/releaseCTAButtons/ReleaseRevertButton'
+import {ReleaseRevertButton} from '../components/releaseCTAButtons/ReleaseRevertButton/ReleaseRevertButton'
 import {ReleaseScheduleButton} from '../components/releaseCTAButtons/ReleaseScheduleButton'
 import {ReleaseUnscheduleButton} from '../components/releaseCTAButtons/ReleaseUnscheduleButton'
 import {ReleaseMenuButton} from '../components/ReleaseMenuButton/ReleaseMenuButton'
